src/requirements.txt
auto_mix_prep==0.2.0
anthropic==0.40.0
dataclasses_json==0.6.7
deepmerge==2.0
firebase_admin==6.6.0
Flask==3.1.0
Flask_Cors==5.0.0
gunicorn==23.0.0
openai==1.57.4
pandas==2.2.3
pydantic==2.10.3
PyPDF2==3.0.1
pytest==8.3.4
PyYAML==6.0.2
PyYAML==6.0.2
Werkzeug==3.1.3
google-cloud-tasks==2.18.0
python-json-logger==3.2.1
redis==5.2.1
google-cloud-storage==2.19.0
google-cloud-secret-manager==2.22.0